# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/events/{eventId}/ticket-types": {
      "post": {
        "operationId": "EventsController_createTicketType",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/v1/events/{eventId}/reserve": {
      "post": {
        "operationId": "EventsController_reserveTickets",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/v1/events/{eventId}/tickets/validate": {
      "post": {
        "operationId": "EventsController_validateTicket",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/v1/events/{eventId}/scanner/validate": {
      "post": {
        "operationId": "ScannerController_validateTicket",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "qrPayload",
                  "signature"
                ],
                "properties": {
                  "qrPayload": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "signature": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Validation result determined.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "VALID"
                    },
                    "ticketId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "ticketType": {
                      "type": "string",
                      "nullable": true
                    },
                    "scannedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed payload."
          },
          "401": {
            "description": "Scanner device unauthenticated."
          },
          "403": {
            "description": "Scanner not authorized for this event."
          }
        },
        "security": [
          {
            "ScannerAuth": []
          }
        ],
        "summary": "Validate a cryptographic ticket payload at the gate",
        "tags": [
          "Events - Gate Scanner"
        ]
      }
    },
    "/api/v1/tickets/wallet": {
      "get": {
        "operationId": "TicketsController_getWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wallet retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tickets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "ticketNumber": {
                            "type": "string"
                          },
                          "event": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "title": {
                                "type": "string"
                              },
                              "startTime": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "venue": {
                                "type": "string"
                              }
                            }
                          },
                          "ticketType": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "credential": {
                            "type": "object",
                            "properties": {
                              "payload": {
                                "type": "object"
                              },
                              "signature": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve the authenticated customer's ticket wallet",
        "tags": [
          "Events - Ticket Wallet"
        ]
      }
    },
...
    "/api/v1/admin/kyc/cases/{id}/approve": {
      "post": {
        "operationId": "AdminKycController_approveCase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminKyc"
        ]
      }
    },
    "/api/v1/admin/kyc/cases/{id}/reject": {
      "post": {
        "operationId": "AdminKycController_rejectCase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminKyc"
        ]
      }
    },
...
  },
  "info": {
    "title": "Animanga Platform API",
    "description": "Strictly defined contracts for Commerce, Events, and Identity",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "ScannerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-scanner-api-key"
      }
    },
...
  }
}